Starting story mode implementation
